Add ITensors as normal dependency
The two package extesions we currently have, ITensorMPSExt and
LindbladVectorizedTensors, depend on ITensors, but their loading is not
triggered by ITensors alone. This means that ITensors is not
automatically available within those extensions.

The closest we can get to what we want with the current functionality
is (see https://discourse.julialang.org/t/using-multiple-packages-in-extensions-for-julia-1-9/98237/5?u=phx):

* make ITensors a normal dependency to our package,
* but only load it in the extensions.

Users will unconditionally install ITensors, but will only pay the
cost of loading it once the extensions load.